refactor: Reorganize Overview component file structure
Move Overview component from components/ to dedicated overview/ directory to
better organize related features and maintain a cleaner component hierarchy
